refactured Unit Test
